+++ src/snowflake/snowpark/session.py
@@ -3,6 +3,7 @@
 # Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
 #
 
+import atexit
 import datetime
 import decimal
 import inspect
@@ -209,6 +210,23 @@ def _add_session(session: "Session") -> None:
         _active_sessions.add(session)
 
 
+def _close_session_atexit():
+    """
+    This is the helper function to close all active sessions at interpreter shutdown. For example, when a jupyter
+    notebook is shutting down, this will also close all active sessions and make sure send all telemetry to the server.
+    """
+    with _session_management_lock:
+        for session in _active_sessions.copy():
+            try:
+                session.close()
+            except Exception:
+                pass
+
+
+# Register _close_session_atexit so it will be called at interpreter shutdown
+atexit.register(_close_session_atexit)
+
+
 def _remove_session(session: "Session") -> None:
     with _session_management_lock:
         try:
@@ -344,7 +362,7 @@ class Session:
                 session = self._create_internal(self._options.get("connection"))
 
             if self._app_name:
-                app_name_tag = f'APPNAME={self._app_name}'
+                app_name_tag = f"APPNAME={self._app_name}"
                 session.append_query_tag(app_name_tag)
 
             return session
@@ -1913,8 +1931,12 @@ class Session:
     def sql(self, query: str, params: Optional[Sequence[Any]] = None) -> DataFrame:
         """
         Returns a new DataFrame representing the results of a SQL query.
-        You can use this method to execute a SQL statement. Note that you still
-        need to call :func:`DataFrame.collect` to execute this query in Snowflake.
+
+        Note:
+            You can use this method to execute a SQL query lazily,
+            which means the SQL is not executed until methods like :func:`DataFrame.collect`
+            or :func:`DataFrame.to_pandas` evaluate the DataFrame.
+            For **immediate execution**, chain the call with the collect method: `session.sql(query).collect()`.
 
         Args:
             query: The SQL statement to execute.