+++ CHANGELOG.md
@@ -9,6 +9,15 @@
 - Added support for the following local testing functions:
   - greatest
   - least
+  - strip_null_value
+  - upper
+  - lower
+  - length
+  - initcap
+
+### Improvements
+
+- Added cleanup logic at interpreter shutdown to close all active sessions.
 
 ### Bug Fixes